+++ brmspy/brms.py
@@ -16,7 +16,7 @@ from brmspy.install import install_brms, install_prebuilt, get_brms_version
 from brmspy.brms_functions.io import get_brms_data, read_rds_fit, read_rds_raw, save_rds
 from brmspy.brms_functions.prior import prior, get_prior, default_prior
 from brmspy.brms_functions.brm import brm, fit
-from brmspy.brms_functions.diagnostics import summary, fixef, ranef, posterior_summary, prior_summary, loo, loo_compare
+from brmspy.brms_functions.diagnostics import summary, fixef, ranef, posterior_summary, prior_summary, loo, loo_compare, validate_newdata
 from brmspy.brms_functions.generic import call
 from brmspy.brms_functions.formula import formula
 from brmspy.brms_functions.prediction import posterior_epred, posterior_linpred, posterior_predict, log_lik
@@ -55,7 +55,7 @@ __all__ = [
     "posterior_predict", "posterior_epred", "posterior_linpred", "log_lik",
 
     # diagnosis
-    'summary', 'fixef', 'ranef', 'posterior_summary', 'prior_summary', 'loo', 'loo_compare',
+    'summary', 'fixef', 'ranef', 'posterior_summary', 'prior_summary', 'loo', 'loo_compare', 'validate_newdata',
     
     # generic
     'call',